+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/joinReorder/JoinReorderSuite.scala
@@ -15,19 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.catalyst.optimizer
+package org.apache.spark.sql.catalyst.optimizer.joinReorder
 
 import org.apache.spark.sql.catalyst.dsl.expressions._
 import org.apache.spark.sql.catalyst.dsl.plans._
 import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeMap}
-import org.apache.spark.sql.catalyst.plans.{Cross, Inner, PlanTest}
+import org.apache.spark.sql.catalyst.optimizer._
+import org.apache.spark.sql.catalyst.plans.{Cross, Inner}
 import org.apache.spark.sql.catalyst.plans.logical.{ColumnStat, LocalRelation, LogicalPlan}
 import org.apache.spark.sql.catalyst.rules.RuleExecutor
 import org.apache.spark.sql.catalyst.statsEstimation.{StatsEstimationTestBase, StatsTestPlan}
 import org.apache.spark.sql.internal.SQLConf.{CBO_ENABLED, JOIN_REORDER_ENABLED}
 
 
-class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
+class JoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestBase {
 
   object Optimize extends RuleExecutor[LogicalPlan] {
     val batches =
@@ -129,7 +130,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
         .select(outputsOf(t1, t2, t3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("put unjoinable item at the end and reorder 3 joinable tables") {
@@ -146,7 +147,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t4)
         .select(outputsOf(t1, t2, t4, t3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("reorder 3 tables with pure-attribute project") {
@@ -161,7 +162,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
         .select(nameToAttr("t1.v-1-10"))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("reorder 3 tables - one of the leaf items is a project") {
@@ -178,7 +179,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           Some(nameToAttr("t1.k-1-2") === nameToAttr("t5.k-1-5")))
         .select(nameToAttr("t1.v-1-10"))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("don't reorder if project contains non-attribute") {
@@ -188,7 +189,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t3, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t3.v-1-100")))
         .select("key".attr)
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, originalPlan)
+    assertEqualPlans(Optimize, originalPlan, originalPlan)
   }
 
   test("reorder 4 tables (bushy tree)") {
@@ -209,7 +210,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t4.k-1-2")))
         .select(outputsOf(t1, t4, t2, t3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("keep the order of attributes in the final output") {
@@ -221,7 +222,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           .join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
           .select(expectedOrder: _*)
       // The plan should not change after optimization
-      assertEqualPlansForJoinReorder(Optimize, expectedPlan, expectedPlan)
+      assertEqualPlans(Optimize, expectedPlan, expectedPlan)
     }
   }
 
@@ -239,7 +240,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           .join(tab2, Cross, Some('b === 'i))
           .select(outputsOf(tab1, tab2, tab3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, original, expected)
+    assertEqualPlans(Optimize, original, expected)
   }
 
   test("reorder recursively") {
@@ -283,7 +284,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
       .union(t4.select(nameToAttr("t4.v-1-10")))
       .join(t5, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t5.v-1-5")))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
   }
 
   test("don't reorder if hints present") {
@@ -296,7 +297,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           Inner,
           Some(nameToAttr("t1.k-1-2") === nameToAttr("t4.k-1-2")))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, originalPlan)
+    assertEqualPlans(Optimize, originalPlan, originalPlan)
 
     val originalPlan2 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -305,7 +306,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .hint("broadcast")
         .join(t3, Inner, Some(nameToAttr("t4.v-1-10") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan2, originalPlan2)
+    assertEqualPlans(Optimize, originalPlan2, originalPlan2)
 
     val originalPlan3 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -313,7 +314,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t3, Inner, Some(nameToAttr("t1.v-1-10") === nameToAttr("t3.v-1-100")))
         .join(t5, Inner, Some(nameToAttr("t5.v-1-5") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan3, originalPlan3)
+    assertEqualPlans(Optimize, originalPlan3, originalPlan3)
   }
 
   test("reorder below and above the hint node") {
@@ -329,7 +330,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
       .select(outputsOf(t1, t2, t3): _*)
       .hint("broadcast").join(t4)
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan, bestPlan)
+    assertEqualPlans(Optimize, originalPlan, bestPlan)
 
     val originalPlan2 =
       t1.join(t2).join(t3)
@@ -343,7 +344,7 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .select(outputsOf(t1, t2, t3): _*)
         .join(t4.hint("broadcast"))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan2, bestPlan2)
+    assertEqualPlans(Optimize, originalPlan2, bestPlan2)
 
     val originalPlan3 =
       t1.join(t2, Inner, Some(nameToAttr("t1.k-1-2") === nameToAttr("t2.k-1-5")))
@@ -360,10 +361,6 @@ class JoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(t4, Inner, Some(nameToAttr("t4.v-1-10") === nameToAttr("t3.v-1-100")))
         .join(t5, Inner, Some(nameToAttr("t5.v-1-5") === nameToAttr("t3.v-1-100")))
 
-    assertEqualPlansForJoinReorder(Optimize, originalPlan3, bestPlan3)
-  }
-
-  private def outputsOf(plans: LogicalPlan*): Seq[Attribute] = {
-    plans.map(_.output).reduce(_ ++ _)
+    assertEqualPlans(Optimize, originalPlan3, bestPlan3)
   }
 }
