+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/joinReorder/StarJoinReorderSuite.scala
@@ -15,18 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.catalyst.optimizer
+package org.apache.spark.sql.catalyst.optimizer.joinReorder
 
 import org.apache.spark.sql.catalyst.dsl.expressions._
 import org.apache.spark.sql.catalyst.dsl.plans._
 import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeMap}
-import org.apache.spark.sql.catalyst.plans.{Inner, PlanTest}
+import org.apache.spark.sql.catalyst.optimizer._
+import org.apache.spark.sql.catalyst.plans.Inner
 import org.apache.spark.sql.catalyst.plans.logical.{ColumnStat, LocalRelation, LogicalPlan}
 import org.apache.spark.sql.catalyst.rules.RuleExecutor
 import org.apache.spark.sql.catalyst.statsEstimation.{StatsEstimationTestBase, StatsTestPlan}
 import org.apache.spark.sql.internal.SQLConf._
 
-class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
+class StarJoinReorderSuite extends JoinReorderPlanTestBase with StatsEstimationTestBase {
 
   var originalConfStarSchemaDetection = false
   var originalConfCBOEnabled = true
@@ -184,7 +185,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d2, f1, d3, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 2: Star join on a subset of dimensions due to inequality joins") {
@@ -223,7 +224,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, f1, d2, s3, d3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 3: Star join on a subset of dimensions since join column is not unique") {
@@ -259,7 +260,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, f1, d2, s3, d3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 4: Star join on a subset of dimensions since join column is nullable") {
@@ -296,7 +297,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") < nameToAttr("s3_pk1")))
         .select(outputsOf(d1, f1, d2, s3, d3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 5: Table stats not available for some of the joined tables") {
@@ -334,7 +335,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
           Some(nameToAttr("f1_fk2") === nameToAttr("d2_pk1")))
         .join(s3, Inner, Some(d3_fk1 === nameToAttr("s3_pk1")))
 
-    assertEqualPlansForJoinReorder(Optimize, query, equivQuery)
+    assertEqualPlans(Optimize, query, equivQuery)
   }
 
   test("Test 6: Join with complex plans") {
@@ -367,7 +368,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(d2.where(nameToAttr("d2_c2") === 2), Inner,
           Some(nameToAttr("f1_fk2") === nameToAttr("d2_pk1")))
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 7: Comparable fact table sizes") {
@@ -403,7 +404,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("f11_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, f11, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, equivQuery)
+    assertEqualPlans(Optimize, query, equivQuery)
   }
 
   test("Test 8: No RI joins") {
@@ -439,7 +440,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d3, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 9: Complex join predicates") {
@@ -475,7 +476,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d3, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 10: Less than two dimensions") {
@@ -510,7 +511,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d3, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 11: Expanding star join") {
@@ -544,7 +545,7 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") < nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d3, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
+    assertEqualPlans(Optimize, query, expected)
   }
 
   test("Test 12: Non selective star join") {
@@ -578,10 +579,6 @@ class StarJoinReorderSuite extends PlanTest with StatsEstimationTestBase {
         .join(s3, Inner, Some(nameToAttr("d3_fk1") === nameToAttr("s3_pk1")))
         .select(outputsOf(d1, d3, f1, d2, s3): _*)
 
-    assertEqualPlansForJoinReorder(Optimize, query, expected)
-  }
-
-  private def outputsOf(plans: LogicalPlan*): Seq[Attribute] = {
-    plans.map(_.output).reduce(_ ++ _)
+    assertEqualPlans(Optimize, query, expected)
   }
 }
